\name{addFeatureData}
\alias{addFeatureData}
\title{Adds chromosomal arm annotation to the featureData}
\description{
Adds chromosomal arm annotation to the featureData
}
\usage{
addFeatureData(snpset)
}
\arguments{
  \item{snpset}{An object inherited from class \code{SnpLevelSet}}
}
\value{
  An \code{AnnotatedDataFrame}
}
\author{R. Scharpf}
\seealso{\code{\link[Biobase]{class.AnnotatedDataFrame}}}
\keyword{manip}

