% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BeadStudioSet}
\alias{BeadStudioSet}
\title{Constructor for BeadStudioSet class}
\usage{
BeadStudioSet(filenames, header_info, genome = c("hg19", "hg18"),
  annotationPkg, chromosome = 1:22, ...)
}
\arguments{
\item{filenames}{path to files containing log R ratios / BAFs}

\item{header_info}{a list of arguments to pass to \code{read.table}, typically created by \code{\link{headerInfo}}}

\item{genome}{either 'hg18' or 'hg19'}

\item{annotationPkg}{character string providing the name of the annotation package}

\item{chromosome}{integer vector indicating which chromosomes to
include in the BeadStudioSet}

\item{...}{ignored}
}
\value{
\code{\linkS4class{BeadStudioSet}}
}
\description{
Constructs an instance of BeadStudioSet from a list of files
containing log R ratios and B allele frequencies.
}
\examples{
path <- system.file("extdata", package="VanillaICE", mustWork=TRUE)
fname <- list.files(path, pattern="genome_studio", full.names=TRUE)
keep <- c("SNP.Name", "Allele1...AB", "Allele2...AB",
          "Log.R.Ratio", "B.Allele.Freq")
labels <- bead_studio_variables()
labels <- setNames(labels, keep[-1])
classes <- c(rep("character", 3), rep("numeric", 2))
header_info <- headerInfo(fname, skip=10, sep=",",
                          keep=keep, labels=labels,
                          classes=classes)
dat <- read_beadstudio(fname, header_info)
pkg <- "human610quadv1bCrlmm"
bsSet <- BeadStudioSet(fname, header_info, "hg18", pkg)
}
\seealso{
\code{\link{read_beadstudio}}, \code{\linkS4class{BeadStudioSet}}, \code{\link{headerInfo}}
}
\keyword{IO,}
\keyword{classes}

