\name{Beta}
\alias{Beta}
\title{Accessor for emission probabilities}
\description{
Accessor for emission probabilities in the hidden Markov model
}
\usage{
Beta(object)
}
\arguments{
  \item{object}{ An object of class \code{HmmParameter}}
}
\value{
  A matrix.  Each column in the matrix corresponds to a sample.
}
\author{R. Scharpf}
\seealso{\code{\link{HmmParameter-class}}}
\keyword{methods}

