\name{bind}
\alias{bind}
\title{Combine two objects of the same class}
\description{
  Combine two objects of the same class. Defined only for classes
  extending \code{hSet}
}
\usage{
bind(object1, object2, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object1}{object  extending class \code{hSet}}
  \item{object2}{must be the same class as \code{object1}}
  \item{\dots}{ Not implemented}
}
\details{

}
\value{
  An object of the same class as \code{object1}
}
\references{}
\author{}
\note{
}
\seealso{ }
\examples{
}
\keyword{methods}

