\name{breakpoints}
\alias{breakpoints}
\title{Locations of breakpoints from the HMM}
\description{
  Returnes indices of SNPs within a chromosome where the breakpoints occur
}
\usage{
breakpoints(object, ...)
}
\arguments{
  \item{object}{object inheriting from class \code{AnnotatedSnpSet}}
  \item{\dots}{additional arguments}
}
\details{
  
}
\value{

  data.frame
  
}
\references{}
\author{}
\note{
}
\seealso{}
\examples{

data(chromosome1)
breakpoints(chromosome1)

}
\keyword{manip}
\keyword{attribute}

