\name{byArm}
\alias{byArm}
\title{Fits the HMM to each chromosomal arm}
\description{
Fits the HMM to each chromosomal arm. Not called directly
}
\usage{
byArm(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{object extending \code{hSet} class}
  \item{\dots}{SCALE: (numeric) factor for scaling transition probability,
    MISSING.CODE for genotype calls, verbose (character string), arm
    (character string) }
}
\details{

}
\value{
  object of the same class as \code{object}
}
\references{}
\author{}
\note{
}
\seealso{}
\examples{
}
\keyword{methods}

