\name{calculateEmissionProbability}
\alias{calculateEmissionProbability}
\title{Calculates emission probabilities for the HMM}
\description{
  
  Calculates emission probabilities for the HMM
  
}
\usage{
calculateEmissionProbability(object, ...)
}

\arguments{
  \item{object}{an object extending \code{hSet}}
  \item{\dots}{optional arguments}
}
\details{

}
\value{

  a matrix of emission probabilities
  
}
\references{}
\author{ }
\note{ 
}
\seealso{}
\examples{
}
\keyword{methods}

