\name{callsIce}
\alias{callsIce}
\alias{callsIce<-}
\title{Indicates whether to use confidence estimates of the genotype
  calls in the HMM}
\description{
  
  When TRUE, the HMM uses the uncertainty estimates of the genotype
  calls (stored in \code{callsConfidence}).
  
}
\usage{
callsIce(object)
}

\arguments{
  \item{object}{ Object of class \code{HmmSnpCallSet}}
}
\details{

}
\value{
  Logical
}
\references{}
\author{}
\note{
}
\seealso{}
\examples{
}
\keyword{methods}

