\name{chromosomeArm}
\alias{chromosomeArm}
\alias{chromosomeArm<-}
\title{Accessor for chromosome arm}
\description{
  Specifies to which chromosome arm(s) the hidden Markov model was fit.
}
\usage{
chromosomeArm(object)
}
\arguments{
  \item{object}{ class inheriting from \code{\link[SNPchip]{AnnotatedSnpSet}}}
}
\details{
}

\value{
  
  A character string specifying the chromosomal arm ("p", "q", or
  "both").
}

\references{}
\author{}
\note{
}
\seealso{}
\examples{
}
\keyword{methods}

