\name{cnEmission-methods}
\docType{methods}
\alias{cnEmission-methods}
\alias{cnEmission,CopyNumberSet,ANY-method}
\alias{cnEmission,matrix,matrix-method}
\alias{cnEmission,oligoSnpSet,ANY-method}
\alias{cnEmission,matrix,matrix-method}
\alias{cnEmission,CopyNumberSet,ANY-method}
\alias{cnEmission,oligoSnpSet,ANY-method}
\title{Methods for Function \code{cnEmission} in Package \pkg{VanillaICE}}
\description{
  Methods for calculating the emission probability for estimates of total copy number.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "CopyNumberSet", stdev = "ANY")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "matrix", stdev = "matrix")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "oligoSnpSet", stdev = "ANY")}}{
%%  ~~describe this method here~~
}
}}

\seealso{\code{\link{cnEmission}}}
\keyword{methods}
