\name{copyNumberIce}
\alias{copyNumberIce}
\alias{copyNumberIce<-}
\title{ Whether to use the uncertainty estimates of the copy number in
  the HMM}
\description{
  When true, the HMM uses the uncertainty estimaes of the copy number
  stored in the \code{assayDataElement} \code{cnConfidence}
}
\usage{
copyNumberIce(object)
}
\arguments{
  \item{object}{Object of class \code{HmmSnpCopyNumberSet}}
}
\details{

}
\value{
  Logical
}
\references{}
\author{ }
\note{
}
\seealso{}
\examples{
}
\keyword{methods}

