% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{HmmOptionList-class}
\alias{HmmOptionList}
\alias{HmmOptionList-class}
\alias{baf,SnpSet2-method}
\alias{hmm.setup}
\title{Defunct functions/classes/methods in the VanillaICE package}
\usage{
HmmOptionList(...)

hmm.setup(...)

\S4method{baf}{SnpSet2}(object)
}
\description{
The function, class, or data object you asked is defunct.
}
\keyword{internal}

