% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{cn_means}
\alias{EMupdates}
\alias{EMupdates,EmissionParam-method}
\alias{EMupdates,HmmParam-method}
\alias{EmissionParam}
\alias{EmissionParam,missing-method}
\alias{EmissionParam,numeric-method}
\alias{baf_means}
\alias{baf_means,EmissionParam-method}
\alias{baf_means,HmmParam-method}
\alias{baf_means<-}
\alias{baf_means<-,EmissionParam,numeric-method}
\alias{baf_sds}
\alias{baf_sds,EmissionParam-method}
\alias{baf_sds,HmmParam-method}
\alias{baf_sds<-}
\alias{baf_sds<-,EmissionParam,numeric-method}
\alias{cn_means}
\alias{cn_means,EmissionParam-method}
\alias{cn_means,HmmParam-method}
\alias{cn_means<-}
\alias{cn_means<-,EmissionParam,numeric-method}
\alias{cn_sds}
\alias{cn_sds,EmissionParam-method}
\alias{cn_sds,HmmParam-method}
\alias{cn_sds<-}
\alias{cn_sds<-,EmissionParam,numeric-method}
\alias{show,EmissionParam-method}
\title{A parameter class for computing Emission probabilities}
\usage{
cn_means(object)

cn_sds(object)

baf_means(object)

baf_sds(object)

baf_means(object) <- value

baf_sds(object) <- value

cn_sds(object) <- value

cn_means(object) <- value

EmissionParam(cn_means = CN_MEANS(), cn_sds = CN_SDS(),
  baf_means = BAF_MEANS(), baf_sds = BAF_SDS(), initial = rep(1/6, 6),
  EMupdates = 5L, CN_range = c(-5, 3), temper = 1, p_outlier = 1/100)

EMupdates(object)

\S4method{show}{EmissionParam}(object)
}
\arguments{
\item{object}{see \code{showMethods("EMupdates")}}

\item{value}{numeric vector}

\item{cn_means}{numeric vector of starting values for log R ratio means (order is by copy number state)}

\item{cn_sds}{numeric vector of starting values for log R ratio standard deviations (order is by copy number state)}

\item{baf_means}{numeric vector of starting values for BAF means ordered.  See example for details on how these are ordered.}

\item{baf_sds}{numeric vector of starting values for BAF means ordered.  See example for details on how these are ordered.}

\item{initial}{numeric vector of intial state probabilities}

\item{EMupdates}{number of EM updates}

\item{CN_range}{the allowable range of log R ratios.  Log R ratios outside this range are thresholded.}

\item{temper}{Emission probabilities can be tempered by emit^temper. This is highly experimental.}

\item{p_outlier}{probability that an observation is an outlier (assumed to be the same for all markers)}
}
\value{
numeric vector
}
\description{
Parameters for computing emission probabilities for a 6-state HMM,
including starting values for the mean and standard deviations for
log R ratios (assumed to be Gaussian) and B allele frequencies
(truncated Gaussian), and initial state probabilities.

Constructor for EmissionParam class

This function is exported primarily for internal use by other BioC
packages.
}
\details{
The log R ratios are assumed to be emitted from a normal
distribution with a mean and standard deviation that depend on the
latent copy number.  Similarly, the BAFs are assumed to be emitted
from a truncated normal distribution with a mean and standard
deviation that depends on the latent number of B alleles relative
to the total number of alleles (A+B).
}
\examples{
ep <- EmissionParam()
cn_means(ep)
ep <- EmissionParam()
cn_sds(ep)
ep <- EmissionParam()
baf_means(ep)
ep <- EmissionParam()
baf_sds(ep)
ep <- EmissionParam()
baf_means(ep) <- baf_means(ep)
ep <- EmissionParam()
baf_sds(ep) <- baf_sds(ep)
ep <- EmissionParam()
cn_sds(ep) <- cn_sds(ep)
ep <- EmissionParam()
cn_means(ep) <- cn_means(ep)
ep <- EmissionParam()
show(ep)
cn_means(ep)
cn_sds(ep)
baf_means(ep)
baf_sds(ep)
}

