\name{emissionProbability}
\alias{emissionProbability}
\alias{emissionProbability<-}
\title{Accessor for the emission probabilities}
\description{
  A matrix of emission probabilities with dimension G x S, where G is
  the number of SNPs and S is the number of hidden states.
}
\usage{
emissionProbability(object)
}
\arguments{
  \item{object}{one of the HMM classes defined in the R package \pkg{ICE}}
}
\details{
}
\value{
  Matrix of emission probabilities
}
\references{

  \cite{RB Scharpf et al. (2007), Department of Biostatistics
    Working Papers, Working Paper 136 }

  }
\author{ }
\note{ 

}
\seealso{}
\examples{
}
\keyword{methods}

