% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filters}
\alias{filters}
\alias{filters,HMM-method}
\alias{filters,HmmParam-method}
\title{Accessor for HMM filter parameters}
\usage{
filters(object)
}
\arguments{
\item{object}{see \code{showMethods(filters)}}
}
\description{
Accessor for HMM filter parameters
}

