\name{fit}
\alias{fit}
\docType{data}
\title{Object of class HmmPredict}
\description{
Object of class \code{HmmPredict}
}
\usage{data(fit)}
\details{
  This object was created from the \code{chromosome1} as described in
  the VanillaICE vignette.
}
\seealso{\code{\link{chromosome1}}}
\examples{
data(fit)
}
\keyword{datasets}
