% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getExampleSnpExperiment}
\alias{getExampleSnpExperiment}
\title{Create an example SnpArrayExperiment from source files containing
marker-level genomic data that are provided in this package}
\usage{
getExampleSnpExperiment()
}
\value{
A \code{\link{SnpArrayExperiment}}
}
\description{
Create an example SnpArrayExperiment from source files containing
marker-level genomic data that are provided in this package
}
\examples{
\dontrun{
   snp_exp <- getExampleSnpExperiment()
}
}

