% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHmmParams}
\alias{getHmmParams}
\alias{getHmmParams,HMM-method}
\alias{getHmmParams,HmmParam-method}
\title{Accessor for HMM model parameters}
\usage{
getHmmParams(object)
}
\arguments{
\item{object}{see \code{showMethods(HmmParam)}}
}
\description{
Accessor for HMM model parameters
}
\examples{
hmm_object <- HMM()
getHmmParams(hmm_object)
}

