\name{getInitialStateProbability}
\alias{getInitialStateProbability}
\title{Initial state probabilities }
\description{
  Retrieves default parameters for the initial state probability distribution
}
\usage{
getInitialStateProbability(object)
}

\arguments{
  \item{object}{ One of the HMM classes defined in \code{ICE}}
}
\details{

}
\value{
a vector of initial state probabilities
}
\references{}
\author{ who you are }
\note{ 
}
\seealso{ }
\examples{
##See defaults
}
\keyword{attribute}

