\name{getLohProbability}
\alias{getLohProbability}
\title{Emission probabilities for LOH}
\description{
  Specifies default emission probabilities for loss of heterozygosity
  (LOH).  We define LOH as a sequence of homozygous genotypes longer
  than what we would expect to observe by chance.
}
\usage{
getLohProbability(object)
}

\arguments{
  \item{object}{ Object of one of the HMM classes defined in \code{ICE}}
}
\details{
   If necessary, more details than the description above 
}
\value{ A matrix of default values for the LOH probability

}
\references{}
\author{}
\note{ 
}
\seealso{}
\examples{
}
\keyword{attribute}

