\name{getPar-methods}
\docType{methods}
\alias{getPar-methods}
\title{Adds data-dependent graphical parameters}
\description{

  Adds graphical parameters to an object inherited from class
      \code{ParHSet} that depend on the data class object (an object
      inherited from class \code{SnpLevelSet}). For instance, graphical
      parameters specifying the layout depend on the number of samples
      and chromosomes in the object \code{snpset}

}

\section{Methods}{
  \describe{
    \item{object = "ParHSet", snpset = "SnpLevelSet"}{Adds graphical
    parameters to the \code{ParHSet} object. }
  }
}
\keyword{methods}

