\name{gtEmission-methods}
\docType{methods}
\alias{gtEmission-methods}
\alias{gtEmission,matrix-method}
\alias{gtEmission,SnpSet-method}
\title{Methods for calculating the emission probabilities of diallelic
  genotypes in Package \pkg{VanillaICE} }
\description{Methods for calculating the emission probabilities of diallelic
  genotypes in Package \pkg{VanillaICE} }
\section{Methods}{
\describe{

\item{\code{signature(object = "matrix")}}{

}

\item{\code{signature(object = "SnpSet")}}{
}
}}
\seealso{\code{\link{gtEmission}}}
\keyword{methods}
