\name{hmmBySample}
\alias{hmmBySample}
\title{Fits HMM to a single sample}
\description{
}
\usage{
hmmBySample(object, ...)
}
\arguments{
  \item{object}{ object extending \code{hSet}}
  \item{\dots}{additional arguments}
}
\details{

}
\value{

  predictions from the HMM for one sample and one chromosome
  
}

\references{}
\author{}
\note{
}
\seealso{}
\examples{
}
\keyword{methods}

