% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{HmmGRanges-class}
\alias{HmmGRanges}
\alias{HmmGRanges,Rle-method}
\alias{HmmGRanges,integer-method}
\alias{HmmGRanges,missing-method}
\alias{HmmGRanges-class}
\alias{state,HmmGRanges-method}
\title{HmmGRanges container}
\usage{
HmmGRanges(states, feature_starts, feature_chrom, loglik,
  emission_param = EmissionParam())

\S4method{state}{HmmGRanges}(object)
}
\arguments{
\item{states}{copy number number state inferred by HMM}

\item{feature_starts}{start location in reference genome [basepairs]}

\item{feature_chrom}{end location in reference genome [basepairs]}

\item{loglik}{the log likelihood}

\item{emission_param}{an instance of \code{EmissionParam} class}

\item{object}{a \code{HmmGRanges} object}
}
\description{
HmmGRanges container
}

