% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{HmmGRanges}
\alias{HmmGRanges}
\title{##HmmGRanges container
##}
\usage{
HmmGRanges(states, feature_starts, feature_chrom, loglik,
  emission_param = EmissionParam())
}
\arguments{
\item{states}{copy number number state inferred by HMM
##}

\item{feature_starts}{start location in reference genome [basepairs]
##}

\item{feature_chrom}{end location in reference genome [basepairs]
##}

\item{loglik}{the log likelihood
##}

\item{emission_param}{an instance of \code{EmissionParam} class
##}
}
\description{
##HmmGRanges container
##
}
\examples{
## library(oligoClasses)
## library(IRanges)
## load(system.file("extdata", "snp_exp.rda", package="VanillaICE"))
## states <- Rle(factor(c(3, 4, 3, 5, 3, 2, 3, 3, 2, 3, 2, 3)),
##               as.integer(c(996, 102, 902, 50, 2467, 102, 76, 1822,
##                            99, 900, 20, 160)))
## hgr <- HmmGRanges(states=states, feature_starts=start(snp_exp),
##                   feature_chrom=chromosome(snp_exp), loglik=15.3)
##
}

