% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{HMMList-class}
\alias{HMMList-class}
\alias{show,HMMList-method}
\alias{unlist,HMMList-method}
\title{Class, constructor, and methods for representing HMM results from
multiple samples}
\usage{
\S4method{show}{HMMList}(object)

\S4method{unlist}{HMMList}(x, recursive = TRUE, use.names = TRUE)
}
\arguments{
\item{object}{a \code{HMMList} object}

\item{x}{a \code{HMMList} object}

\item{recursive}{logical;  currently ignored}

\item{use.names}{logical;  currently ignored}
}
\description{
Each element of the HMMList contains the genomic intervals of the
HMM segmentation (GRanges-derived object), parameters from the
Baum-Welch, a matrix of posterior probabilities for the 6-state HMM
(each row corresponds to a segment), and a \code{FilterParam}
object.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a list. Each element of the list should be a \code{HMM} object.}
}}
\seealso{
\code{\linkS4class{HMM}}
}

