\name{initialStateProbability}
\alias{initialStateProbability}
\alias{initialStateProbability<-}
\title{Retrieves initial state probabilities}

\description{
  
  Accessor for initial state probabilities used to fit the HMM.  The
  replacement method allows the user to specify the initial state
  probability matrix for fitting the HMM. 
  
}

\usage{
initialStateProbability(object)
}
\arguments{
  \item{object}{An object of one of the HMM classes defined in \code{ICE}}
}
\details{

}
\value{
  Matrix of initial state probabilities
}

\references{ }

\author{ }
\note{ 
}
\seealso{}
\examples{
}
\keyword{attribute}
\keyword{methods}

