% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{fileName}
\alias{bafFile}
\alias{bafFile,ArrayViews-method}
\alias{fileName}
\alias{fileName,ArrayViews-method}
\alias{gtFile}
\alias{gtFile,ArrayViews-method}
\alias{lrrFile}
\alias{lrrFile,ArrayViews-method}
\title{Accessor for parsed file(s) containing low level summaries for copy number analysis}
\usage{
\S4method{fileName}{ArrayViews}(object, label)

lrrFile(object, label = "lrr")

bafFile(object, label = "baf")

gtFile(object, label = "gt")

\S4method{lrrFile}{ArrayViews}(object, label = "lrr")

\S4method{bafFile}{ArrayViews}(object, label = "baf")

\S4method{gtFile}{ArrayViews}(object, label = "gt")
}
\arguments{
\item{object}{An array views-type container}

\item{label}{character string to be incorporated as part of the file name for parsed BAFs}
}
\description{
These accessors are not meant to be called directly by the user.  They
are exported in the package NAMESPACE for internal use by other BioC
packages.
}
\examples{
views <- ArrayViews(parsedPath=tempdir())
sourcePaths(views)
lrrFile(views)
bafFile(views)
gtFile(views)
}

