\name{isMissing}
\alias{isMissing}
\title{Checks for missing genotype calls}
\description{

  Checks for missing genotype calls
  
}

\usage{
isMissing(object, ...)
}


\arguments{
  
  \item{object}{an object extending \code{hSet}}
  \item{\dots}{set MISSING.CODE to the value of missingness }
  
}

\details{
  
  
}

\value{

}

\references{}
\author{}
\note{ 
}
\seealso{}
\examples{
}
\keyword{methods}

