\name{likelihood}
\alias{likelihood}
\alias{likelihood<-}
\title{The partial likelihood}

\description{ Accessor defined for the partial likelihood calculated
  from the Viterbi algorithm }

\usage{
likelihood(object)
}
\arguments{
  \item{object}{An object of one of the HMM classes defined in \code{ICE}}
}
\details{
}
\value{
  A numerical vector of the same length as there are SNPs in the object.
}

\references{RB Scharpf et al. (2007), Department of Biostatistics
  Working Papers, Working Paper 136}

\author{}
\note{ 
}
\seealso{}
\examples{
}
\keyword{methods}

