\name{locationCopyNumber}
\alias{locationCopyNumber}
\alias{locationCopyNumber<-}
\title{Location parameters for the copy number distribution (Gaussian)}
\description{ Accessor for the location parameter of the Gaussian
  distribution used to estimate the emission probabilities in the HMM.
}
\usage{
locationCopyNumber(object)
}
\arguments{
  \item{object}{ One of the HMM classes defined in \code{ICE}}
}
\details{
}
\value{ Numerical vector
}
\references{}
\author{ }
\note{ 
}
\seealso{\code{\link{scaleCopyNumber}}}
\examples{
}
\keyword{methods}

