% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{LogLik-class}
\alias{LogLik-class}
\alias{length,LogLik-method}
\alias{show,LogLik-method}
\title{Classes and methods for storing/getting log-likelihoods from Viterbi algorithm}
\usage{
\S4method{length}{LogLik}(x)

\S4method{show}{LogLik}(object)
}
\arguments{
\item{x}{object of class \code{LogLik}}

\item{object}{a \code{LogLik} object}
}
\description{
Exported for internal use by other BioC packages
}
\section{Slots}{

\describe{
\item{\code{loglik}}{a numeric vector}

\item{\code{tolerance}}{a numeric vector}
}}
\seealso{
\code{\link{LogLik}}
}

