\name{numberStates}
\alias{numberStates}
\alias{numberStates<-}
\title{Number of hidden states}
\description{
  Accessor for the number of hidden states=
}
\usage{
numberStates(object)
}
\arguments{
  \item{object}{An object of one of the HMM classes defined in \code{ICE}}
}
\details{
}
\value{
  An integer
}
\references{ }
\author{}
\note{ 
}
\seealso{ }
\examples{
}
\keyword{methods}

