\name{ParHSet-class}
\docType{class}
\alias{ParHSet-class}
\alias{initialize,ParHSet-method}
\title{Class ParHSet}
\description{Graphical parameters}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("ParHSet", border.prediction, legend.prediction, legend.location.prediction, ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{snpPar}:}{Object of class \code{"list"}}
  }
}
\section{Extends}{
Class \code{\link[SNPchip]{ParESet-class}}, directly.
}
\section{Methods}{
  \describe{
    \item{initialize}{\code{signature(.Object = "ParHSet")}}
  }
}
\references{}
\author{}
\note{}

\seealso{
}
\examples{
showClass("ParHSet")
}
\keyword{classes}
