% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{parsedPath}
\alias{parsedPath}
\alias{parsedPath,ArrayViews-method}
\title{Complete path to directory for keeping parsed files}
\usage{
parsedPath(object)

\S4method{parsedPath}{ArrayViews}(object)
}
\arguments{
\item{object}{a \code{ArrayViews} object}
}
\description{
A character string indicating the complete path for storing parsed
files.
}
\seealso{
\code{\link{parseSourceFile}} \code{\linkS4class{ArrayViews}}

\code{\linkS4class{ArrayViews}}
}

