\name{pi}
\alias{pi}
\title{Accessor for the initial state probabilities of the hidden Markov
model}
\description{
Accessor for the initial state probabilities of the hidden Markov
model
}
\usage{
pi(object)
}
\arguments{
  \item{object}{object of class \code{HmmParameter}}
}

\value{
  Numeric vector with length equal to the number of hidden states
}
\author{RS}
\seealso{ \code{\link{HmmParameter-class}}}
\keyword{methods}

