\name{plotPredictions}
\alias{plotPredictions}
\title{Plots the predicted states versus physical position}
\description{
Plots the predicted states versus physical position
}
\usage{
plotPredictions(object, op, breaks, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{an object derived from class \code{hSet}}
  \item{op}{parameter to par and layout}
  \item{breaks}{locations of breakpoints predicted from the HMM}
  \item{\dots}{additional arguments to the initialization methods of the
  classes extending \code{ParHSet}}
}
\details{
}
\value{
  
}
\references{}
\author{}
\note{ 
}
\seealso{ }
\examples{
}
\keyword{aplot}

