% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{xyplotList}
\alias{xyplotList}
\alias{xyplotList,GRangesList,SnpArrayExperiment-method}
\alias{xyplotList,HmmGRanges,SnpArrayExperiment-method}
\title{Lattice-style plots for granges and SnpArrayExperiment objects}
\usage{
xyplotList(granges, se, param = HmmTrellisParam())

\S4method{xyplotList}{HmmGRanges,SnpArrayExperiment}(granges, se,
  param = HmmTrellisParam())

\S4method{xyplotList}{GRangesList,SnpArrayExperiment}(granges, se,
  param = HmmTrellisParam())
}
\arguments{
\item{granges}{a \code{HmmGRanges} object}

\item{se}{a \code{SnpArrayExperiment}}

\item{param}{trellis parameters for plotting HMM}
}
\description{
Data for the graphic is generated by a call to \code{grangesData}.
}

