\name{predictions}
\alias{predictions}
\alias{predictions<-}
\title{Predicted states from the HMM}
\description{
  Vector of predicted states from the HMM
}
\usage{
predictions(object)
}
\arguments{
  \item{object}{ One of the HMM classes defined in \code{ICE}}
}
\details{
}
\value{
  Vector of integers
}
\references{}
\author{}
\note{ 
}
\seealso{}
\examples{
}
\keyword{methods}

