% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.bsfiles}
\alias{read.bsfiles}
\title{Wrapper for reading GenomeStudio files}
\usage{
read.bsfiles(...)
}
\arguments{
\item{...}{arguments passed to fread}
}
\description{
Wrapper for reading GenomeStudio files
}
\examples{
path <- system.file("extdata", package="VanillaICE", mustWork=TRUE)
file <- list.files(path, pattern="genome_studio", full.names=TRUE)
## get the headers
headers <- names(read_bsfiles(file, nrows=0))
select <- match(c("SNP Name", "Allele1 - AB", "Allele2 - AB", "Log R Ratio", "B Allele Freq"), headers)
dat <- read.bsfiles(file, select=select)
}

