% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rescale}
\alias{rescale}
\title{Rescale a numeric vector}
\usage{
rescale(x, l, u)
}
\arguments{
\item{x}{numeric vector}

\item{l}{lower limit of rescaled \code{x}}

\item{u}{upper limit of rescaled \code{x}}
}
\description{
Rescale a numeric vector
}
\keyword{manip}

