% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rowMAD}
\alias{robustSds}
\alias{rowMAD}
\title{Compute the median absolute deviation (MAD) for the rows of a matrix}
\usage{
rowMAD(x, ...)

robustSds(x, takeLog = FALSE, ...)
}
\arguments{
\item{x}{matrix}

\item{...}{additional arguments to rowMedians}

\item{x}{a numeric matrix}

\item{takeLog}{whether to first log2 transform the numeric matrix}

\item{...}{additional arguments to \code{mad}}
}
\value{
a matrix of the same dimension as the input. Within a column, the entries are identical
}
\description{
Compute the median absolute deviation (MAD) for the rows of a matrix

Calculate the column-wise median absolute deviation of a matrix
}
\examples{
X <- matrix(rnorm(100), 10, 10)
rowMedians(X)
data(locusLevelData, package="oligoClasses")
sds <- robustSds(locusLevelData[["copynumber"]]/100,
                 takeLog=TRUE)
}
\seealso{
\code{\link{mad}}

\code{\link{mad}} \code{\link{rowMedians}}

\code{\link{mad}}
}

