\name{scaleCopyNumber}
\alias{scaleCopyNumber}
\alias{scaleCopyNumber<-}
\title{Variance parameters for the  copy number distribution (Gaussian)}
\description{ Accessor for the variance parameter of the Gaussian
  distribution used to estimate the emission probabilities in the HMM.
}

\usage{
scaleCopyNumber(object)
}
\arguments{
  \item{object}{ One of the HMM classes defined in \code{ICE}}
}
\details{
}
\value{ Numeric vector
}
\references{ }
\author{ }
\note{
}
\seealso{\code{\link{locationCopyNumber}}}
\examples{
}
\keyword{methods}

