% Generated by roxygen2 (4.0.2): do not edit by hand
\name{snpArrayAssays}
\alias{snpArrayAssays}
\title{Create an assays object from log R ratios and B allele frequencies}
\usage{
snpArrayAssays(cn = new("matrix"), baf = new("matrix"), ...)
}
\arguments{
\item{cn}{matrix of log R ratios}

\item{baf}{matrix of B allele frequencies}

\item{...}{one can pass additional matrices to the SnpExperiment
constructor, such as the SNP genotypes.}
}
\description{
Create an assays object from log R ratios and B allele frequencies
}

