% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{SnpArrayExperiment-class}
\alias{SnpArrayExperiment}
\alias{SnpArrayExperiment-class}
\title{SnpArrayExperiment container for log R ratios and B allele frequencies}
\usage{
SnpArrayExperiment(cn, baf, rowData = GRanges(), colData = DataFrame(),
  isSnp = logical(), ...)
}
\arguments{
\item{cn}{matrix of copy number estimates (e.g., log R ratios)}

\item{baf}{matrix of B allele frequencies}

\item{rowData}{GRanges object for SNPs/nonpolymorphic markers}

\item{colData}{DataFrame containing sample-level covariates}

\item{isSnp}{logical vector indicating whether marker is a SNP}
}
\description{
Constructor for SnpArrayExperiment
}

