% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{SnpExperiment}
\alias{SnpExperiment}
\alias{SnpExperiment,ArrayViews-method}
\title{Constructor for SnpArrayExperiment}
\usage{
SnpExperiment(object)

\S4method{SnpExperiment}{ArrayViews}(object)
}
\arguments{
\item{object}{see \code{showMethods('SnpExperiment')} for a list of supported objects}
}
\description{
A single-argument generic function to construct a SnpArrayExperiment.
}
\examples{
view <- ArrayViews()
SnpExperiment(view)

## create a views object from GenomeStudio source files
  require(BSgenome.Hsapiens.UCSC.hg18)
  require(data.table)
  require(foreach)
  extdir <- system.file("extdata", package="VanillaICE", mustWork=TRUE)

  features <- suppressWarnings(fread(file.path(extdir, "SNP_info.csv")))
  fgr <- GRanges(paste0("chr", features$Chr), IRanges(features$Position, width=1),
                 isSnp=features[["Intensity Only"]]==0)
  fgr <- SnpGRanges(fgr)
  names(fgr) <- features[["Name"]]
  seqlevels(fgr) <- seqlevels(BSgenome.Hsapiens.UCSC.hg18)[seqlevels(BSgenome.Hsapiens.UCSC.hg18) \%in\% seqlevels(fgr)]
  seqinfo(fgr) <- seqinfo(BSgenome.Hsapiens.UCSC.hg18)[seqlevels(fgr),]
  fgr <- sort(fgr)
  files <- list.files(extdir, full.names=TRUE, recursive=TRUE, pattern="FinalReport")
  ids <- gsub(".rds", "", gsub("FinalReport", "", basename(files)))
  dat <- fread(files[1], nrows=0)
  keep <- c("SNP Name", "Allele1 - AB", "Allele2 - AB", "Log R Ratio", "B Allele Freq")
  select <- which(names(dat)\%in\%keep)
  dat <- fread(files[1], select=select)
  index_genome <- match(names(fgr), dat[["SNP Name"]])
  scan_params <- CopyNumScanParams(index_genome=index_genome,
                                   select=as.integer(select))
  views <- ArrayViews(rowData=fgr,
                      sourcePaths=files,
                      sample_ids=ids,
                      parsedPath=tempdir())
  sapply(views, parseSourceFile, scan_params)
  se <- SnpExperiment(views)
}

