% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SnpGRanges-class}
\alias{SnpGRanges}
\alias{SnpGRanges,GRanges-method}
\alias{SnpGRanges,missing-method}
\alias{SnpGRanges-class}
\title{An extension to GRanges for representing SNPs}
\usage{
SnpGRanges(object = GRanges(), isSnp, ...)
}
\arguments{
\item{object}{A \code{GRanges} object}

\item{isSnp}{A logical vector.  Each genomic interval in the
\code{GRanges} container corresponds to a marker on the genotyping
array.  \code{isSnp} is FALSE for nonpolymorphic markers such as
those included on the Affymetrix 6.0 chips.}

\item{...}{ignored}
}
\description{
An extension to GRanges for representing SNPs

Constructor for SnpGRanges class
}
\section{Slots}{

\describe{
\item{\code{elementMetadata}}{a \code{SnpDataFrame}}
}}
\examples{
SnpGRanges()
}

