% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{start,oligoSnpSet-method}
\alias{start,oligoSnpSet-method}
\title{Retrieve genomic location of SNPs}
\usage{
\S4method{start}{oligoSnpSet}(x)
}
\arguments{
\item{x}{a univariate or multivariate time-series, or a vector or matrix.}
}
\description{
Retrieve genomic location of SNPs
}

