\name{stateNames}
\alias{stateNames}
\alias{stateNames<-}
\title{The hidden states}

\description{
  Accessor for the names of the hidden states
}

\usage{
stateNames(object)
}

\arguments{
  \item{object}{ One of the HMM classes defined in \code{ICE}}
}
\details{
}
\value{ A character vector
}
\references{}
\author{}
\note{
}
\seealso{}
\examples{
}
\keyword{methods}

