\name{states}
\alias{states}
\alias{states<-}
\title{Accessor and replacement method for the hidden states of the
  Markov model}
\description{
  The latent states for the hidden Markov model.
}
\usage{
        states(object)
        states(object) <- value
}

\arguments{
  \item{object}{An object of class \code{HmmOptions} or
    \code{HmmParameter}}
  \item{value}{Vector of class \code{character}.}
}
\value{
  Character string  of hidden states
}
\author{RS}
\keyword{methods}
