% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sweepMode}
\alias{sweepMode}
\alias{sweepMode,SnpArrayExperiment-method}
\title{Sweep the modal log R ratio (row-wise) from a matrix of log R ratios}
\usage{
sweepMode(x, MARGIN)
}
\arguments{
\item{x}{an object containing log R ratios}

\item{MARGIN}{integer indicating which margin (1=rows, 2=columns)
to sweep the mode}
}
\description{
This function is most useful when a large number of samples (more
than 10) are available and the dataset is a collection of germline
samples.  We assume that the samples are from a single batch and
that the modal value will be a robust estimate of the mean log R
ratio for diploid copy number.  Variation in the modal estimates
between markers is presumed to be attributable to probe effects
(e.g., differences hybridization efficiency/PCR do to sequence
composition). For sex chromosomes, one should apply this function
separately to men and women and then recenter the resulting matrix
according to the expected copy number.
}

