\name{Viterbi-methods}
\docType{methods}
\alias{Viterbi-methods}
\alias{emission}
\alias{emission,Vit-method}
\title{Methods for Viterbi objects}
\description{
  Methods for Viterbi objects
}
\section{Methods}{

 In the following methods, \code{object} is of class \code{Viterbi} or
 \code{Viterbi2}.

  \describe{

    \item{}{\code{emission(object)}: Accessor for the emission
      probabilities.}
  }

}
\keyword{methods}

