% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xygrid}
\alias{xygrid}
\title{Layout the lattice-style plots for the CNV data on a grid}
\usage{
xygrid(trellis_plot, viewports, granges)
}
\arguments{
\item{trellis_plot}{an object of class \code{trellis}}

\item{viewports}{a list of viewports as provided by the \code{viewports} function}

\item{granges}{a  \code{HmmGRanges} object}
}
\description{
Layout the lattice-style plots for the CNV data on a grid
}
\seealso{
\code{\link{viewports}}
}

