\name{cnEmission}
\alias{cnEmission}
\title{Calculate emission probabilities for copy number}
\description{
Calculate emission probabilities for the copy number
}
\usage{
cnEmission(x, options, ...)
}
\arguments{
  \item{x}{(log) copy number (vector)}
  \item{options}{ an object of class \code{HmmOptions}}
  \item{\dots}{Not implemented}
}
\value{
  Vector of the logged emission probabilites
}
\author{R. Scharpf}
\keyword{methods}

