% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{filters}
\alias{filters}
\alias{filters,HmmParam-method}
\alias{filters,HMM-method}
\title{Accessor for HMM filter parameters}
\usage{
filters(object)
}
\arguments{
\item{object}{see \code{showMethods(filters)}}
}
\description{
Accessor for HMM filter parameters
}
