% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SnpArrayExperiment.R
\name{getExampleSnpExperiment}
\alias{getExampleSnpExperiment}
\title{Create an example SnpArrayExperiment from source files containing
marker-level genomic data that are provided in this package}
\usage{
getExampleSnpExperiment(bsgenome)
}
\arguments{
\item{bsgenome}{a \code{BSgenome} object}
}
\value{
A \code{\link{SnpArrayExperiment}}
}
\description{
Create an example SnpArrayExperiment from source files containing
marker-level genomic data that are provided in this package
}
\examples{
\dontrun{
   if(require("BSgenome.Hsapiens.UCSC.hg18")){
     genome <- BSgenome.Hsapiens.UCSC.hg18
     snp_exp <- getExampleSnpExperiment(genome)
   }
}
}

