% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{HMM-class}
\alias{HMM}
\alias{HMM-class}
\alias{show,HMM-method}
\alias{state,HMM-method}
\title{Container for the segmented data and the 6-state HMM model parameters}
\usage{
HMM(granges = GRanges(), param = HmmParam(), posterior = matrix(),
  filters = FilterParam())

\S4method{state}{HMM}(object)

\S4method{show}{HMM}(object)
}
\arguments{
\item{granges}{a \code{GRanges} object}

\item{param}{a \code{HmmParam} object}

\item{posterior}{matrix of posterior probabilities}

\item{filters}{an object of class \code{FilterParam}}

\item{object}{a \code{HMM} object}
}
\description{
Container for the segmented data and the 6-state HMM model parameters
}
\section{Slots}{

\describe{
\item{\code{granges}}{a \code{GRanges} object}

\item{\code{param}}{a \code{HmmParam} object}

\item{\code{posterior}}{a matrix of posterior probabilities}

\item{\code{filters}}{a \code{FilterParam} object}
}}
\seealso{
\code{\link{hmm2}}
}

