\name{hmmBySample}
\alias{hmmBySample}
\title{Fits HMM to a single sample}

\description{
  Fits the HMM to each chromosomal arm in a sample -- this method is
  generally not called directly by the user.  Instead, one should use
  the hmm method that basically loops over hmmBySample.
}

\usage{
hmmBySample(object, ...)
}

\arguments{
  \item{object}{ object extending \code{hSet}}
  \item{\dots}{additional arguments}
}
\details{

}
\value{

  predictions from the HMM for one sample and one chromosome
  
}

\references{}
\author{Robert B. Scharpf}
\note{
}
\seealso{}
\examples{
}
\keyword{methods}

