% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{HmmGRanges-class}
\alias{HmmGRanges}
\alias{HmmGRanges-class}
\title{HmmGRanges container}
\usage{
HmmGRanges(states, feature_starts, feature_chrom, loglik)
}
\arguments{
\item{states}{copy number number state inferred by HMM}

\item{feature_starts}{start location in reference genome [basepairs]}

\item{feature_chrom}{end location in reference genome [basepairs]}

\item{loglik}{the log likelihood}
}
\description{
HmmGRanges container
}

