% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{hmmResults}
\alias{hmmResults}
\title{Example output from the hidden markov model}
\format{a \code{GRanges} object}
\description{
The results of a 6-state HMM fit to simulated copy number and genotype data.
}
\keyword{datasets}

