% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.bsfiles}
\alias{read.bsfiles}
\title{Deprecated function in VanillaICE}
\usage{
read.bsfiles(...)
}
\description{
The function read.bsfiles  has been deprecated.  This function
is provided only for compatability with older versions and will be
defunct at the next release.
}
\details{
Use the replacement function fread instead.
}
\keyword{internal}

