% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rowModes}
\alias{rowModes}
\alias{colModes}
\alias{rowMAD}
\title{Robust statistics for matrices}
\usage{
rowModes(x)

colModes(x)

rowMAD(x, ...)
}
\arguments{
\item{x}{matrix}

\item{...}{additional arguments to rowMedians}
}
\value{
numeric vector
}
\description{
Compute the column-wide or row-wise mode of numeric matrices
}
\examples{
X <- matrix(rnorm(100), 10, 10)
rowMAD(X)
}
\seealso{
\code{\link{mad}}

\code{\link{mad}} \code{\link{rowMedians}}
}
