% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SnpArrayExperiment-class}
\alias{SnpArrayExperiment}
\alias{SnpArrayExperiment,matrix-method}
\alias{SnpArrayExperiment,missing-method}
\alias{SnpArrayExperiment-class}
\title{A SummarizedExperiment-derived class of marker-level SNP array data
for copy number inference}
\usage{
SnpArrayExperiment(cn, baf, rowData = GRanges(), colData = DataFrame(),
  isSnp = logical(), ...)

\S4method{SnpArrayExperiment}{missing}(cn, baf, rowData = GRanges(),
  colData = DataFrame(), isSnp = logical(), ...)

\S4method{SnpArrayExperiment}{matrix}(cn, baf, rowData = GRanges(),
  colData = DataFrame(row.names = colnames(cn)), isSnp = logical(), ...)
}
\arguments{
\item{cn}{matrix of copy number estimates (e.g., log R ratios)}

\item{baf}{matrix of B allele frequencies}

\item{rowData}{GRanges object for SNPs/nonpolymorphic markers}

\item{colData}{DataFrame containing sample-level covariates}

\item{isSnp}{logical vector indicating whether marker is a SNP}

\item{...}{additional arguments passed to initialization method for \code{SummarizedExperiment}}
}
\description{
A SummarizedExperiment-derived class of marker-level SNP array data
for copy number inference

Constructor for SnpArrayExperiment
}
\examples{
SnpArrayExperiment()
data(snp_exp) # example
SnpArrayExperiment(cn=lrr(snp_exp), baf=baf(snp_exp),
                   rowData=rowData(snp_exp))
}

