% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{start,oligoSnpSet-method}
\alias{start,oligoSnpSet-method}
\title{Retrieve genomic location of SNPs}
\usage{
\S4method{start}{oligoSnpSet}(x)
}
\arguments{
\item{x}{a \code{oligoSnpSet} object}
}
\description{
Retrieve genomic location of SNPs
}

