\name{transitionProbability}
\alias{transitionProbability}
\alias{transitionProbability<-}
\title{Transition probabilities in the HMM}
\description{
  Accessor for the transition probabilities in the HMM
}
\usage{
transitionProbability(object)
}
\arguments{
  \item{object}{ One of the HMM classes defined in \pkg{ICE}}
}
\details{
}
\value{ A matrix
}
\references{ }
\author{ }
\note{ 
}
\seealso{\code{\link{distance}}}
\examples{
}
\keyword{methods}

