% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{VanillaICE}
\alias{VanillaICE}
\alias{VanillaICE-package}
\title{A hidden markov model for detection of germline copy number variants from arrays}
\description{
A hidden markov model for detection of germline copy number variants from arrays
}

