% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{VanillaICE}
\alias{VanillaICE}
\title{A hidden markov model for detection of germline copy number
  variants from arrays}
\description{A hidden markov model for detection of germline copy number
  variants from arrays}
