\name{viterbi}
\alias{viterbi}
\title{viterbi algorithm}

\description{

  The Viterbi algorithm for computing the most likely state sequence
  given a model

}
\usage{
viterbi(object, ...)
}

\arguments{
  \item{object}{object extending \code{hSet}}
  \item{\dots}{optional}
}
\details{

}
\value{
  vector of integers
}
\references{ }
\author{}
\note{ 
}
\seealso{}
\examples{
}
\keyword{methods}

