\name{viterbi}
\alias{viterbi}
\title{viterbi algorithm}
\description{

}
\usage{
viterbi(object, ...)
}

\arguments{
  \item{object}{object extending \code{hSet}}
  \item{\dots}{optional}
}
\details{

}
\value{
  vector of integers
}
\references{ }
\author{}
\note{ 
}
\seealso{}
\examples{
}
\keyword{methods}

